3.3-permissions/api_with_restrictions/requirements.txt
django~=4.1.3
djangorestframework~=3.14.0
psycopg2-binary
django-filter~=22.1
django-debug-toolbar~=3.7.0